Reorganize the project to get the macros as a workspace lib
