+++ .github/workflows/ci.yaml
@@ -31,6 +31,9 @@ jobs:
       - uses: actions/checkout@v4
       - uses: Swatinem/rust-cache@v2
 
+      - name: Run doc tests
+        run: cargo test --doc --workspace --features sqlite
+
       - name: Run lib tests on sqlite
         run: cargo test --lib --features sqlite
