+++ .vscode
@@ -1,0 +1,3 @@
+{
+    "rust-analyzer.cargo.features": ["sqlite"]
+}
