+++ tiny-orm-macros/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "tiny-orm-macros"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2.0", features = ["full", "extra-traits"] }
+quote = "1.0"
+proc-macro2 = "1.0"
+convert_case = "<1.0"
+
+[features]
+default = []
+sqlite = ["sqlx/sqlite"]
+postgres = ["sqlx/postgres"]
+mysql = ["sqlx/mysql"]
+
+[dev-dependencies]
+sqlx = { version = "0.8.2", features = ["runtime-tokio-native-tls", "sqlite", "chrono", "uuid", "derive"] }
+tiny-orm-macros = { path = "."}
+tokio = { version = "1.41.0", features = ["rt", "macros"]}
+uuid = {version = "1.11.0", features = ["v4"]}
