+++ tiny-orm-macros/src/types.rs
@@ -1,5 +1,4 @@
 use convert_case::{Case, Casing};
-use proc_macro2::Span;
 use quote::format_ident;
 use std::{fmt, str::FromStr};
 use syn::{Ident, Type};
@@ -136,10 +135,6 @@ pub struct TableName(pub String);
 impl TableName {
     pub fn new(input: String) -> Self {
         Self(input.to_case(Case::Snake))
-    }
-
-    pub fn to_ident(&self) -> Ident {
-        Ident::new(&self.0, Span::call_site())
     }
 }
